requirements.txt
lightgbm==2.3.1
neptune-client==0.4.130
neptune-contrib==0.27.0
numpy==1.19.0
scikit-learn==0.23.1
scikit-plot==0.3.7
mysql-connector-python==8.0.23   
mysqlclient==2.0.3
